requirements.txt
boto==2.44.0
bz2file==0.98
funcsigs==1.0.2
gensim==0.13.3
mock==2.0.0
numpy==1.11.3
pbr==1.10.0
protobuf==3.2.0
requests==2.12.3
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
sklearn==0.0
smart-open==1.3.5
tensorflow==1.0.1
wheel==0.24.0
annoy
pyyaml